moved Data and Axis reltaed objects to a dedicated data module
