+++ src/pymodaq/examples/function_plotter.py
@@ -9,6 +9,7 @@ import numpy as np
 from qtpy import QtWidgets
 from qtpy.QtCore import Slot, QDate, QThread, QTimer
 
+from pymodaq.utils import data as data_mod
 from pymodaq.utils.logger import set_logger, get_module_name
 from pymodaq.utils.gui_utils.custom_app import CustomApp
 from pymodaq.utils.gui_utils.dock import DockArea, Dock
@@ -112,8 +113,8 @@ class FunctionPlotter(CustomApp):
 
         function_vals = eval(f'np.{function_str}')
 
-        self.viewer.show_data([function_vals], labels=[function_str], x_axis=utils.Axis(x, label='An axis',
-                                                                                        units='arb. units'))
+        self.viewer.show_data([function_vals], labels=[function_str], x_axis=data_mod.Axis(x, label='An axis',
+                                                                                             units='arb. units'))
 
     def value_changed(self, param):
         if param.name() == 'add_function':
