+++ src/pymodaq/extensions/daq_logger.py
@@ -10,6 +10,7 @@ import sys
 from collections import OrderedDict
 import datetime
 
+import utils.data
 from pymodaq.utils.logger import set_logger, get_module_name
 from pymodaq.utils.gui_utils.custom_app import CustomApp
 from pymodaq.utils.gui_utils.dock import Dock
@@ -18,7 +19,7 @@ import pymodaq.utils.parameter.ioxml
 import numpy as np
 from qtpy import QtWidgets
 from qtpy.QtCore import QObject, Slot, QThread, Signal, Qt
-
+from pymodaq.utils import data as data_mod
 from pymodaq.utils.gui_utils.widgets import QLED
 from pymodaq.utils import daq_utils as utils
 from pymodaq.utils.h5modules import H5Logger
@@ -435,8 +436,8 @@ class DAQ_Logging(QObject):
     def format_actuators_data(self, act_name, act_value):
         acq_time = datetime.datetime.now().timestamp()
         data = OrderedDict(name=act_name, acq_time_s=acq_time, control_module='DAQ_Move',
-                           data0D=OrderedDict(data=utils.DataToExport(name=act_name, dim='Data0D', source='raw',
-                                                                          data=np.array([act_value]))))
+                           data0D=OrderedDict(data=data_mod.DataToExport(name=act_name, dim='Data0D', source='raw',
+                                                                           data=np.array([act_value]))))
         self.do_save_continuous(data)
 
     def connect_actuators(self, connect=True):
