+++ tests/utils_test/daq_utils_test.py
@@ -1,6 +1,7 @@
 import numpy as np
 import os
 
+from pymodaq.utils import data as data_mod
 from pymodaq.utils.parameter import utils as putils
 import utils.units
 from pymodaq.utils.logger import set_logger, get_module_name
@@ -178,36 +179,36 @@ def test_ThreadCommand():
 
 
 def test_Axis():
-    ax = utils.Axis()
+    ax = data_mod.Axis()
     assert 'data' in ax
     assert 'label' in ax
     assert 'units' in ax
 
     assert ax.label == ax['label']
 
-    ax = utils.Axis(np.array([1, 2, 3, 5, 7]), 'a label', 'seconds')
+    ax = data_mod.Axis(np.array([1, 2, 3, 5, 7]), 'a label', 'seconds')
     assert np.all(ax['data'] == np.array([1, 2, 3, 5, 7]))
     assert ax['label'] == 'a label'
     assert ax['units'] == 'seconds'
 
-    ax = utils.Axis(label=None, units=None)
+    ax = data_mod.Axis(label=None, units=None)
     assert ax['label'] == ''
     assert ax['units'] == ''
 
     with pytest.raises(TypeError):
-        utils.Axis(10)
+        data_mod.Axis(10)
     with pytest.raises(TypeError):
-        utils.Axis(label=10)
+        data_mod.Axis(label=10)
     with pytest.raises(TypeError):
-        utils.Axis(units=10)
+        data_mod.Axis(units=10)
 
 
 def test_NavAxis():
-    navaxis_tmp = utils.NavAxis(nav_index=1)
-    assert isinstance(navaxis_tmp, utils.NavAxis)
+    navaxis_tmp = data_mod.NavAxis(nav_index=1)
+    assert isinstance(navaxis_tmp, data_mod.NavAxis)
     assert navaxis_tmp['nav_index'] == 1
     with pytest.raises(ValueError):
-        utils.NavAxis()
+        data_mod.NavAxis()
 
 
 class TestData:
@@ -215,107 +216,107 @@ class TestData:
         name = 'data_test'
         x = utils.linspace_step(1, 100, 1)
         y = utils.linspace_step(0.01, 1, 0.01)
-        data_test = utils.Data(name=name, x_axis=x, y_axis=y)
-        assert isinstance(data_test, utils.Data)
+        data_test = data_mod.Data(name=name, x_axis=x, y_axis=y)
+        assert isinstance(data_test, data_mod.Data)
         assert data_test['name'] == name
-        assert data_test['x_axis'] == utils.Axis(data=x)
-        assert data_test['y_axis'] == utils.Axis(data=y)
+        assert data_test['x_axis'] == data_mod.Axis(data=x)
+        assert data_test['y_axis'] == data_mod.Axis(data=y)
 
-        x = utils.Axis(x)
-        y = utils.Axis(y)
+        x = data_mod.Axis(x)
+        y = data_mod.Axis(y)
         kwargs = [1, 2.0, 'kwargs', True, None]
-        data_test = utils.Data(name=name, x_axis=x, y_axis=y, kwargs=kwargs)
+        data_test = data_mod.Data(name=name, x_axis=x, y_axis=y, kwargs=kwargs)
         assert data_test['x_axis'] == x
         assert data_test['y_axis'] == y
         assert data_test['kwargs'] == kwargs
 
         with pytest.raises(TypeError):
-            utils.Data(name=None)
+            data_mod.Data(name=None)
         with pytest.raises(TypeError):
-            utils.Data(source=None)
+            data_mod.Data(source=None)
         with pytest.raises(ValueError):
-            utils.Data(source='source')
+            data_mod.Data(source='source')
 
         with pytest.raises(TypeError):
-            utils.Data(distribution=None)
+            data_mod.Data(distribution=None)
         with pytest.raises(ValueError):
-            utils.Data(distribution='distribution')
+            data_mod.Data(distribution='distribution')
 
         with pytest.raises(TypeError):
-            utils.Data(x_axis=10)
+            data_mod.Data(x_axis=10)
         with pytest.raises(TypeError):
-            utils.Data(y_axis=10)
+            data_mod.Data(y_axis=10)
 
     def test_DataFromPlugins(self):
         data = [utils.linspace_step(1, 100, 1), utils.linspace_step(0.01, 1, 0.01)]
         nav_axes = ["test"]
-        x_axis = utils.Axis(data=utils.linspace_step(1, 100, 1))
-        y_axis = utils.Axis(data=utils.linspace_step(1, 100, 1))
-        data_test = utils.DataFromPlugins(data=data, nav_axes=nav_axes, nav_x_axis=x_axis, nav_y_axis=y_axis)
-        assert isinstance(data_test, utils.DataFromPlugins)
+        x_axis = data_mod.Axis(data=utils.linspace_step(1, 100, 1))
+        y_axis = data_mod.Axis(data=utils.linspace_step(1, 100, 1))
+        data_test = data_mod.DataFromPlugins(data=data, nav_axes=nav_axes, nav_x_axis=x_axis, nav_y_axis=y_axis)
+        assert isinstance(data_test, data_mod.DataFromPlugins)
         assert data_test['data'] == data
         assert data_test['nav_axes'] == nav_axes
         assert data_test['nav_x_axis'] == x_axis
         assert data_test['nav_y_axis'] == y_axis
         assert data_test['dim'] == 'Data1D'
         data = [np.array([1])]
-        data_test = utils.DataFromPlugins(data=data)
+        data_test = data_mod.DataFromPlugins(data=data)
         assert data_test['dim'] == 'Data0D'
         data = [np.array([[1, 1], [1, 2]])]
-        data_test = utils.DataFromPlugins(data=data)
+        data_test = data_mod.DataFromPlugins(data=data)
         assert data_test['dim'] == 'Data2D'
         data = [np.array([[[1, 1], [1, 2]], [[2, 1], [2, 2]]])]
-        data_test = utils.DataFromPlugins(data=data)
+        data_test = data_mod.DataFromPlugins(data=data)
         assert data_test['dim'] == 'DataND'
 
         with pytest.raises(TypeError):
-            utils.DataFromPlugins(data=[1, 2, 3, 4, 5])
+            data_mod.DataFromPlugins(data=[1, 2, 3, 4, 5])
         with pytest.raises(TypeError):
-            utils.DataFromPlugins(data="str")
+            data_mod.DataFromPlugins(data="str")
 
     def test_DataToExport(self):
         data = np.array([1])
-        data_test = utils.DataToExport(data=data)
-        assert isinstance(data_test, utils.DataToExport)
+        data_test = data_mod.DataToExport(data=data)
+        assert isinstance(data_test, data_mod.DataToExport)
         assert data_test['data'] == data
         assert data_test['dim'] == 'Data0D'
-        data_test = utils.DataToExport()
+        data_test = data_mod.DataToExport()
         assert data_test['dim'] == 'Data0D'
         data = np.array([1, 1])
-        data_test = utils.DataToExport(data=data)
+        data_test = data_mod.DataToExport(data=data)
         assert data_test['dim'] == 'Data1D'
         data = np.array([[1, 1], [1, 2]])
-        data_test = utils.DataToExport(data=data)
+        data_test = data_mod.DataToExport(data=data)
         assert data_test['dim'] == 'Data2D'
         data = np.array([[[1, 1], [1, 2]], [[2, 1], [2, 2]]])
-        data_test = utils.DataToExport(data=data)
+        data_test = data_mod.DataToExport(data=data)
         assert data_test['dim'] == 'DataND'
 
         with pytest.raises(TypeError):
-            utils.DataToExport(data="data")
+            data_mod.DataToExport(data="data")
 
 
 def test_ScaledAxis():
-    scaled_axis = utils.ScaledAxis()
-    assert isinstance(scaled_axis, utils.ScaledAxis)
+    scaled_axis = data_mod.ScaledAxis()
+    assert isinstance(scaled_axis, data_mod.ScaledAxis)
     assert scaled_axis['offset'] == 0
     assert scaled_axis['scaling'] == 1
 
     with pytest.raises(TypeError):
-        utils.ScaledAxis(offset=None)
+        data_mod.ScaledAxis(offset=None)
     with pytest.raises(TypeError):
-        utils.ScaledAxis(scaling=None)
+        data_mod.ScaledAxis(scaling=None)
     with pytest.raises(ValueError):
-        utils.ScaledAxis(scaling=0)
+        data_mod.ScaledAxis(scaling=0)
 
     assert scaled_axis['scaling'] == scaled_axis.scaling
 
 
 def test_ScalingOptions():
-    scaling_options = utils.ScalingOptions(utils.ScaledAxis(), utils.ScaledAxis())
-    assert isinstance(scaling_options, utils.ScalingOptions)
-    assert isinstance(scaling_options['scaled_xaxis'], utils.ScaledAxis)
-    assert isinstance(scaling_options['scaled_yaxis'], utils.ScaledAxis)
+    scaling_options = data_mod.ScalingOptions(data_mod.ScaledAxis(), data_mod.ScaledAxis())
+    assert isinstance(scaling_options, data_mod.ScalingOptions)
+    assert isinstance(scaling_options['scaled_xaxis'], data_mod.ScaledAxis)
+    assert isinstance(scaling_options['scaled_yaxis'], data_mod.ScaledAxis)
 
 
 def test_recursive_find_files_extension():
